RenamedComp
